Document Shell

+++ src/os/user/Shell/Commands/Rm.cc
@@ -1,3 +1,20 @@
+/*
+ * Copyright (C) 2018 Fabian Ruhland
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 #include "Rm.h"
 
 Rm::Rm(Shell &shell) : Command(shell) {
@@ -5,6 +22,11 @@ Rm::Rm(Shell &shell) : Command(shell) {
 };
 
 void Rm::deleteFile(const String &progName, FileStatus &fStat) {
+    if(shell.getCurrentWorkingDirectory().getAbsolutePath().beginsWith(fStat.getAbsolutePath())) {
+        stderr << progName << ": '" << fStat.getName() << "': Cannot delete the current working directory!" << endl;
+        return;
+    }
+
     auto ret = fileSystem->deleteFile(fStat.getAbsolutePath());
 
     switch(ret) {
